backend/requirements.txt
PyGithub==2.5.0
python-dotenv==1.0.1
manim>=0.19.0
